BDOG-260 Refactor Forms.notEmpty
